Filter network metrics endpoint
